adding admin delete endpoint and related tests & optimized import

+++ apps/todoapp/src/routers/todos.py
@@ -8,9 +8,9 @@ from fastapi import status
 from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
 
+from .auth import get_current_user
 from ..database import SessionLocal
 from ..models import Todos
-from .auth import get_current_user
 
 router = APIRouter()
 
@@ -86,16 +86,13 @@ async def update_todo(user: user_dependency,
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(user: user_dependency,
-                      db: db_dependency,
-                      todo_id: int = Path(gt=0)):
-    if user is None or user.get('user_role') != 'admin':
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
         raise HTTPException(status_code=401, detail='Authentication Failed')
 
-    todo_model = (db.query(Todos).filter(Todos.id == todo_id)
-                  .filter(Todos.owner_id == user.get('id'))
-                  .first())
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
